Extract shared TypeScript/zod rendering into baklava-tscommon
The tsrest and orpc formatters carried near-identical copies of the zod
schema renderer, union collapsing, TS escaping/identifier quoting,
multipart body rendering, param-group building, and the path->name /
camelCase derivations. One more TS-emitting format would have made it
three copies drifting independently.

New `tscommon` module: TsZodRenderer parameterized by TsZodDialect (the
few places the formats legitimately differ — date-time schema, file
part schema, empty-union fallback) plus TsNaming. Both formatters
delegate; their specs and the gold files are untouched — the gold suite
passes WITHOUT regeneration on both Scala versions, proving the
refactor changes no output byte.

tsfetch is left alone: it renders plain TS types, not zod, and shares
nothing but trivia.
