+++ .github/workflows/ci.yml
@@ -75,11 +75,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p scalatest/target http4sroutes/target http4s/target postman/target openapi/target tsfetch/target pekkohttproutes/target simple/target orpc/target sbtplugin/target sttpclient/target munit/target core/target tsrest/target specs2/target pekkohttp/target project/target
+        run: mkdir -p scalatest/target http4sroutes/target http4s/target postman/target openapi/target tsfetch/target pekkohttproutes/target simple/target orpc/target sbtplugin/target sttpclient/target munit/target core/target tsrest/target specs2/target pekkohttp/target tscommon/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar scalatest/target http4sroutes/target http4s/target postman/target openapi/target tsfetch/target pekkohttproutes/target simple/target orpc/target sbtplugin/target sttpclient/target munit/target core/target tsrest/target specs2/target pekkohttp/target project/target
+        run: tar cf targets.tar scalatest/target http4sroutes/target http4s/target postman/target openapi/target tsfetch/target pekkohttproutes/target simple/target orpc/target sbtplugin/target sttpclient/target munit/target core/target tsrest/target specs2/target pekkohttp/target tscommon/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
